Support Bazel build system (#1033)
Bug:https://github.com/ArthurSonzogni/FTXUI/issues/1032
Fixed:https://github.com/ArthurSonzogni/FTXUI/issues/1032

+++ src/ftxui/component/menu_test.cpp
@@ -266,7 +266,7 @@ TEST(MenuTest, MenuEntryIndex) {
   menu->OnEvent(Event::ArrowDown);
   menu->OnEvent(Event::ArrowDown);
   menu->OnEvent(Event::Return);
-  for (int index = 0; index < menu->ChildCount(); index++) {
+  for (size_t index = 0; index < menu->ChildCount(); index++) {
     EXPECT_EQ(menu->ChildAt(index)->Index(), index);
   }
 }
